requirements.txt
gradio==4.20.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3